Major changes
The database approach is changed from standard
to use content provider. Also there are added
groups so the user can organize the tasks in groups.
Added management for the groups and tasks. Also
there is a basic implementation for reminders for the
tasks, but approack is so bad, and has to be changed
(probably in the next commit).